nn: Add test partitioning runtime tests.
These were disabled due to a dependency on HardSwish, which requires a
TFLite uprev. Temporarily disable these tests until the uprev occurs so
that we can compile and link the test files, and run the other tests.

For TestPartitioningRandom this effectively disables all of the test,
but it's an improvement to have it compiled and linked.

BUG=b:157948826
TEST=FEATURES=test emerge-betty chromeos-base/aosp-frameworks-ml-nn

Change-Id: I5c416de7ed4869a14b1b803155875ea4d2ed36c7
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/frameworks/ml/+/2256105
Commit-Queue: Alan Green <avg@chromium.org>
Tested-by: Stuart Langley <slangley@chromium.org>
Auto-Submit: Stuart Langley <slangley@chromium.org>
Reviewed-by: Alan Green <avg@chromium.org>