nn: Use dynamic linking for sample driver.
Rather than statically linking, use dynamic linking for sample driver so
that it is loaded the same way that a vendor nn hal library will be
loaded.

To support this move the inclusion of "chromeos/sampledriver.cpp" to be
linked as part of SampleDriver and not runtime.

Fix missing symbols for the sample driver shared library.

Followup work: Look at making common a shared library so we don't need
to link it statically to neuralnetworks.so and vendor-nn-hal.so, to save
on space. It might be that vendors would like to re-use this
functionality as well.

BUG=b:155721846
TEST=Deploy to device, run ml_cmdline --nnapi successfully.

Change-Id: I3b31fdf29fbffdd77290c21c20eeef8beda39ec6
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/frameworks/ml/+/2256117
Tested-by: Stuart Langley <slangley@chromium.org>
Auto-Submit: Stuart Langley <slangley@chromium.org>
Reviewed-by: Jim Pollock <jmpollock@chromium.org>
Commit-Queue: Jim Pollock <jmpollock@chromium.org>