nn: fix segfault due to static initialisation race
test_wrapper::Model is being statically initialised, which eventually calls
DeviceManager::findAvailableDevices which also depends on a statically
initialised object (IDevice::getService). This isn't necessary, so
convert the model initialisation to be done in SetUp; there wasn't a
good reason for it not to be.

Enable TestMemoryDomain, but disable the tests that call into hardware
buffers which are not yet implemented.

BUG=b:159414899
TEST=FEATURES=test emerge-kukui chromeos-base/aosp-frameworks-ml-nn

Change-Id: I6e13686f3fa56191ff6a7313793126356323a937
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/frameworks/ml/+/2258172
Tested-by: Jim Pollock <jmpollock@chromium.org>
Commit-Queue: Jim Pollock <jmpollock@chromium.org>
Auto-Submit: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Stuart Langley <slangley@chromium.org>