nn: ensures all the generated tests are loaded
Previously, TestGenerated ran a set of tests from runtime/test/generated
but that set was produced as a side effect of
TestRemoveDefaultArguments. Within the latter, it loads some of the
object files from the generated test static lib, but the others are
never loaded and hence never run.

Linking with --whole-archive is a way to ensure all the object files are
loaded (as done in Android.bp[whole_static_libs]). This is difficult to
do cleanly in GN.

This CL adds a dummy test that will load all the object files and thus,
the tests. A handful are commented out that are not working, next steps
are to fix those.

BUG=b:159414785
TEST=FEATURES=test emerge-betty chromeos-base/aosp-frameworks-ml-nn

Change-Id: I169b67026b0f8cfd2cea9cf6132f6b945ceb58a9
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/frameworks/ml/+/2260599
Tested-by: Jim Pollock <jmpollock@chromium.org>
Auto-Submit: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Stuart Langley <slangley@chromium.org>
Commit-Queue: Jim Pollock <jmpollock@chromium.org>