philox_random.h: use stdint instead of stdlib
Replace include of stdlib with include of stdint.  philox_random uses
the standard types uint{16,32,64}_t. These are defined in stdint.h.
Android's stdlib.h includes stdint.h, allowing this file to compile,
however this is an include-what-you-use error.

This change is necessary to allow philox_random.h to compiled on
ChromiumOS. In the ChromiumOS toolchsain, stdlib.h does not include
stdint.h

Test: Still compiles

Change-Id: Icbdf6d7e2107aa9cff73769f257a0ed2595b8056