nn: reduce default log verbosity for tests
Currently a large amount of ERROR severity log messages are printed, but
are expected due to the nature of the tests. package_test doesn't
facilitate propagate env variables (ANDROID_LOG_TAGS), so we set it in
the testrunner as a workaround.

BUG=b:162318479
TEST=FEATURES=test emerge-betty aosp-frameworks-ml-nn

Change-Id: I8b7e6ffed0ad32d417b3fe3faf9539732a886a43
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/frameworks/ml/+/2323636
Tested-by: Jim Pollock <jmpollock@chromium.org>
Auto-Submit: Jim Pollock <jmpollock@chromium.org>
Commit-Queue: Stuart Langley <slangley@chromium.org>
Reviewed-by: Stuart Langley <slangley@chromium.org>