Fix integer overflow in RGG.
Some RGG tests crashed when enabling the integer_overflow sanitizer
because the code attempted to take the negative of a uint32_t value
before casting to int32_t. This CL fixed the issue.

Without the sanitizer, the code will behave correctly, so the fix will
not alter the test behavior.

Test: NNAPI CTS
Test: NeuralNetworksTest_static_fuzzing with integer_overflow sanitizer
Change-Id: Id9ba1752ca063b44c78abbbd71e611b1d0f84fdf