Update NNAPI canonical validation
Previously, there was a gap in our validation code when it came to
validating operations because it relied on code that was not compatible
with the canonical types at the time. Now that the code is compatible,
this CL closes the validation gap.

To facilitate this, this CL also relocates some utility code
(HalVersion, NN_RET_CHECK_*) from Utils.h to TypeUtils.h.

Bug: 160668435
Test: mma
Change-Id: I4b9ed9ad14f79ad84860809f72c0ccbd6399e6f0