Disable NNAPI host build
A host build was enabled for canonical types and tests to ensure that
the canonical types were not introducing new dependencies. Because we
are not focusing on developing a host build for the short term, this CL
disables it, but it's being tracked in b/146324523.

Bug: N/A
Test: mma
Change-Id: Ib96349d17b202b163c4e377515b9c3117cc39b1d
Merged-In: Ib96349d17b202b163c4e377515b9c3117cc39b1d
(cherry picked from commit bc56b4b2dbc5b6c9b992a174afc211cc762c3895)