Validate during NN conversions by default -- runtime
The sibling change in this topic renames all `convert` functions to
`unvalidatedConvert`. That change also introduces new `convert`
functions that act only on the types that appear in the NN HIDL methods
directly. These new `convert` functions perform validation.
Specifically, if either the source or destination value is invalid, then
the conversion fails.

In response to the sibling change, this change instead calls
`unvalidatedConvert` for those types that do not have a `convert`
function.

Bug: 160667419
Test: mma
Test: NeuralNetworksTest_static
Change-Id: I12b30ca9c728c64cadb063661c49eb97f8f94745