Remove extensions allowlist for non-system clients on Android S.
NNAPI extensions are restricted to allowlisted non-system clients.
We do not allow any system dependency on extensions, as requested by Treble.
If there's a functionality that system client needs, it has to be in the API.
Allowlisting was self-imposed to prevent NNAPI fragmentation. We received
feedback from Vendors, that this is a requirement for deprecating their custom solutions for accessing
ML accelerators.

This change removes non-system clients allowlist on S devices.

Bug: 177310149
Test: NNT_static
Change-Id: Iebf8152f4aa4ca9d877c78a951c9e9dc303d8b3c