nn: fix race condition in execution tests
The setup of the execution tests is designed to synchronize a call to
getOutputOperandDimensions whilst an execution is running. The atomics
that were controlling this were letting that call happen before the
execution had started, and in rare (high CPU load) occasions this
would deadlock.

This CL fixes the checks on those atomics to ensure the synchronization
happens as expected. A change to configureExecutionBurst was also
required in order to get the SampleDriver to call the wrapped model
(and hence the dummyExecution), otherwise the locking semaphore wouldn't
clear and the test would freeze.

Bug: 163081732
Test: NeuralNetworksTest_static on Pixel 3 (blueline)
Change-Id: Id78d2b25f84607f9e5fe8eb50efe11e2f94edb5f