Passing padding information to the driver -- runtime.
This CL changes the NDK specification and canonical type to enable
passing padding information of execution inputs and outputs to drivers.

The padded length from NDK is split into raw length and padding in
the canonical DataLocation struct. This makes it easy to convert the
canonical Request to HIDL Request where only raw length is expected.

Bug: 179691454
Test: NNT_static
Test: CtsNNAPITestCases
Change-Id: I2bb16f981e1cd1f6b84a05b76311ed5abdd48d5e