Change the behavior of NNAPI burst
This change includes the following two behavior changes:

First, this removes the "preferPowerOverLatency" argument from
"configureExecutionBurst". This is being removed in preparation for
switching the runtime over to using the canonical IBurst interface,
whose "configureExecutionBurst" method does not include a
"preferPowerOverLatency" argument as it serves no purpose on newer HAL
versions. Implementations of the IBurst interface can instead directly
rely on the ExecutionPreference value used to prepare the
IPreparedModel object from which the IBurst was created.

Second, it changes the default burst controller polling time from 50us
to 0us. The previous behavior was to use a polling time of 0us when
"preferPowerOverLatency" was true, and to use a default of 50us when
"preferPowerOverLatency" was false. However, because
"preferPowerOverLatency" is being removed, the default is always set to
0us to account for the preference to use lower power.

Bug: 180492058
Test: mma
Test: NeuralNetworksTest_static
Change-Id: I73417d0f4250d6a8df6260de70f03f94afe77a56
Merged-In: I73417d0f4250d6a8df6260de70f03f94afe77a56
(cherry picked from commit 33f4a8f89ac75b34258426aa1ae96654b5842907)