Guard all usages of ExecutionBuilder::mState by mutex.
This CL changes ExecutionBuilder::mState to use a non-atomic type, and
guards all usages by a mutex.

Prior to this CL, only the writes to ExecutionBuilder::mState are
guarded by a mutex. This is to reduce the total number of lock/unlock
constructing an ANeuralNetworksExecution object. However, profiling
shows no noticeable performance difference before and after this change.

Bug: 183240785
Test: NNT_static
Change-Id: I934142c5e402e1da9fd260316b84935b5aaa44f4