Create first verison of the NNAPI updatable shell service
This first version hardcodes some values such as the service metadata
(number of cache files, capabilities, etc.) as well as the full support
library binary path.

This CL includes a fallback "InvalidDevice" for when the support library
is not able to be located, opened, or used. To test this functionality,
modify the name of the prebuilt path to an invalid path or change the
function name "ANeuralNetworks_getSLDriverImpl" to something else.

Bug: 170696191
Test: VtsHalNeuralnetworksTargetTest
Change-Id: I882e520b33d527ab1a1e212e2c24b08df3723994