nn: use cryptohome daemon storage location for cache files
This CL forces any request that sets up compilation caching to store
the files in the cryptohome daemon storage location of the ml_service
process. This is done due to the docs indicating that cache files should
be stored in a secure location. See:
https://source.android.com/devices/neural-networks/compilation-caching#security

We can't do this in configuration since the daemon storage is per-user
and the username hash needs to be queried at runtime.

BUG=b:178762478
TEST=FEATURES=test emerge-volteer aosp-frameworks-ml-nn

Change-Id: I09d5d4be01a3435cedda4e822a4f423f8723e5ce
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/frameworks/ml/+/2817759
Tested-by: Jim Pollock <jmpollock@chromium.org>
Auto-Submit: Jim Pollock <jmpollock@chromium.org>
Commit-Queue: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Stuart Langley <slangley@chromium.org>
Reviewed-by: Michael Pishchagin <mblsha@google.com>