Add ANNShimRegistrationParams to libneuralnetworks_shim.so
This CL extends libneuralnetworks_shim.so with more parameters to
have more control over how a service is registered, including whether
it is a lazy service and whether an invalid device should be
registered in the event the intended device failed to register.

Specifically, this CL introduces the following new llndk functions to
libneuralnetworks_shim.so:
* ANeuralNetworksShimRegistrationParams_create
* ANeuralNetworksShimRegistrationParams_free
* ANeuralNetworksShimRegistrationParams_addDeviceInfo
* ANeuralNetworksShimRegistrationParams_setNumberOfListenerThreads
* ANeuralNetworksShimRegistrationParams_registerAsLazyService
* ANeuralNetworksShimRegistrationParams_fallbackToMinimumSupportDevice

This CL additionally replaces
ANeuralNetworksShim_registerSupportLibraryService's arguments with
"registrationParams" to make the arguments more configurable in future
releases.

Bug: 170696939
Test: mma
Test: NeuralNetworksTest_static
Test: VtsHalNeuralnetworksTargetTest
Change-Id: I6c177490eb055d871386727cd786f0dc0b781893