Change NN partitioner's tiebreaking selection
Prior to this CL, if there were serveral non-updatable IDevice objects
in a list with the same performance characteristics, the partitioner
would select the first device in the list. However, if there were
serveral *updatable* IDevice objects in a list with the same performance
characteristics, the partitioner would instead select the *last* device
in the list.

This CL changes the partitioner's tiebreaking selection behavior to
instead select the first updatable IDevice object in a list.

Bug: N/A
Test: mma
Test: NeuralNetworksTest_static
Change-Id: I215b7f0f8d86052bb973b2a916120c9bfe08f91d