Remove redundant libneuralnetworks_shim calls
The NNAPI shim device needs to know metadata including:
1) How many cache files the driver needs
2) What the performance capabilities of the driver are
3) What vendor extensions the driver supports

Prior to this CL, this information was provided by the shim service or
the shell service through libneuralnetworks_shim. However, this
information is being relocated to the NNAPI support library API, making
the functions in libneuralnetworks_shim redundant.

This CL removes the following API calls:
* ANeuralNetworksShimDeviceInfo_setCapabilities
* ANeuralNetworksShimDeviceInfo_setNumberOfCacheFilesNeeded
* ANeuralNetworksShimDeviceInfo_addVendorExtension

This CL removes the following structs:
* ANeuralNetworksShimPerformanceInfo
* ANeuralNetworksShimOperandPerformanceInfo
* ANeuralNetworksShimExtensionOperandTypeInformation

Bug: 185837243
Test: VtsHalNeuralnetworksTargetTest
Test: NeuralNetworksTest_static
Change-Id: Ifbf1db05bdae67c3e9ff7b2260201f455a8babd7