Change OperandValues' AlignedData to uint8_t
Prior to this CL, Model::OperandValues ensured the data was aligned by
using a vector of aligned data behind the scenes. Specifically, the
AlignedData was std::max_align_t, ensuring it was at least as aligned as
any primitive type used in the NNAPI. This CL changes the underlying
data to use a vector of uint8_t which is simpler to manage, but ensures
the data is still aligned by verifying __STDCPP_DEFAULT_NEW_ALIGNMENT__
is sufficient.

Bug: N/A
Test: mma
Test: NeuralNetworksTest_static
Change-Id: Idd5951f0caa7cae9b2ee990caa8cfd03c9176248