nn: enable NDEBUG, but unset for tests
This CL allows the NDEBUG flag to be set for the core libraries, which
otherwise results in a linking error when using refptr due to the
destructor of RefCountedThreadSafeBase being defined differently based
on the NDEBUG flag. See b/174605330#comment13 for details.

We still require NDEBUG to be unset for tests, so this CL will unset the
define for the test targets, otherwise tests will not build or pass
correctly.

BUG=b:186996014
TEST=FEATURES=test USE=-cros-debug emerge-volteer aosp-frameworks-ml-nn

Change-Id: Icfc69dcd08f96577cc3ca30b0c05df556bfb0550
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/frameworks/ml/+/2870225
Tested-by: Jim Pollock <jmpollock@chromium.org>
Auto-Submit: Jim Pollock <jmpollock@chromium.org>
Commit-Queue: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Stuart Langley <slangley@chromium.org>