Fix generated test initialization order
Prior to this change, the fields in the generated test structures would
be sorted by their name. However, C++20 orders the fields by the order
in the structure's definition. This causes checks like
"reorder-init-list" to fail. This CL fixes the ordering by:
(1) removing the sorted ordering in the test generator
(2) requiring Python 3.7 which preserves ordering in dicts

This change additionally
* fixes the test generator order of the "numberOfConsumers" field
* adds "NOLINT" to supress "using namespace" clang-tidy lint

Bug: 143902513
Test: mma
Test: generate_all_tests.sh
Test: NeuralNetworksTest_static
Change-Id: Icc54d5aeea6ebe9fa0807e27e654e68881e6fb9c