Remove redundant validation from validateRequestForModel
Prior to this CL, validateRequestForModel would validate the model,
validate the request, and validate that the model and request are
compatible (have the same number of inputs, outputs, etc.). However,
validating the model is likely redundant because it should have already
been validated during IDevice::prepareModel*.

This CL removes the redundant model validation and adds a comment to
validateRequestForModel stating that the model is assumed to be valid.

Bug: N/A
Test: mma
Test: NeuralNetworksTest_static
Change-Id: I60fe4d99951eaafe4cd4ea65fc3e015ffbef6650