Skip TestGpuNnapi if the device does not have enough GPU memory.
This CL modifies TestGpuNnapi to choose a heap that is large enough for
memory allocation and skip the test if none of them is suitable. Even if
we found a large enough heap, the device memory allocation may still
fail if the allocation size exceeds the upper limit of a single
allocation that the platform supports. In such a case, the test will
also be skipped.

This CL also fixes a bug when checking the supported extensions --
strcmp returns 0 on match and non-0 on mismatch.

Fixes: 190527027
Test: NNT_static
Change-Id: Iee829307a8cd95ea8701f23cbff279e21266aeb6