Attempt full CPU fallback if a step model has no inputs or no outputs.
Prior to this CL, a step model with zero input or output will be sent to
the driver for compilation. Depending on the error status returned from
the driver, the compilation will fail immediately on INVALID_ARGUMENT,
or result in full CPU fallback on GENERAL_FAILURE.

This CL changes this behavior -- the runtime will avoid sending a step
model with empty input or output to the driver and will always attempt
full CPU fallback if allowed.

Fixes: 192640448
Bug: 192103170
Test: NNT_static
Change-Id: I036596ce53b89cc485c475e73b0f09f9599da059