Change GeneratedTestUtils from AHWB to ASharedMemory
Prior to this CL, GeneratedTestUtils used AHardwareBuffer-backed memory
for TestModel's CONSTANT_REFERENCE memory. This inadvertantly makes it
so any TestModel that is logically V1_0 or V1_1 is changed to version
V1_2 (when AHardwareBuffer-based memory was introduced).

This CL changes GeneratedTestUtils to instead use ASharedMemory via
ANeuralNetworksMemory_createFromFd (which was introduced in V1_0). It
corresponding changes the mobilenet test cases from V1_2 to V1_0 because
of this change.

Bug: 194129896
Test: mma
Test: NeuralNetworksTest_static
Test: CtsNNAPITestCases
Change-Id: I8af4bafe766cfcdc6a66897d84beb6677bff0ff6