Add NNAPI telemetry to Support Library feature level 5
This CL deletes SupportLibraryDiagnostic.h and relocates its functions
to NeuralNetworksSupportLibraryImpl.h. It further removes the _Nonnull
and _Nullable attributes to be consistent with the rest of
NeuralNetworksSupportLibraryImpl.h.

This CL additionally removes the "deviceCount", "packageName",
"fallbackToCpu", and "introspection" fields of the
Diagnostic*Info structs. "deviceCount" can be calculated from
"deviceId". "fallbackToCpu" and "introspection" always return false and
true, respectively. "packageName" always returned the hardcoded string
"<package info not available>".

Bug: 179260976
Test: mma
Change-Id: Iccb6112ac149f7f32a11c6c05f8d826b0d47f4dd