Add newer operations to NNAPI android_fuzzing
This CL adds the following new operations:
  QUANTIZED_LSTM, IF, WHILE, ELU, HARD_SWISH, FILL, RANK,
  BATCH_MATMUL, PACK, MIRROR_PAD, REVERSE

This CL makes a backward-incompatible change to android_fuzzing,
changing the Model message to include control flow operations. This CL
also changes the corresponding conversion code.

Finally, this CL adds an optimization to the Buffer message which allows
it to store empty, a uint32_t scalar type, or a uint32_t seed that can
be used to generate data.

Bug: 206671491
Bug: 148605565
Test: mma
Test: libneuralnetworks_fuzzer
Test: libneuralnetworks_driver_fuzzer
Change-Id: I63830f71a9047c7c9243eabaa957573f06d14694