Modify NnApiSupportLibrary to handle all variants of SL Driver.
With NnApiSupportLibrary being derived from latest version of
NnApiSLDriverImpl, client has to be careful not to call
methods not available in underlying impl - which will result
in dereferencing arbitrary memory, not even the nullptr.

Switched NnApiSupportLibrary to be an union of all NnApiSLDriverImpl
implementations, with getters that assert correctness of getting
the right functions structure.

Test: build SLTM and MLTS
Change-Id: I2f3204d98cc8ef13ae80329485c934f711cd6183