Update OWNERS file
This CL isolates the NNAPI owners into a unique file so that other
OWNERS files can "include" this file, making it simpler to maintain
who the current NNAPI owners are. This CL includes NNAPI_OWNERS from
packages/modules/NeuralNetworks, and a follow up CL will make the same
change for the OWNERS file in hardware/interfaces/neuralnetworks.

This change also:
* adds galarragas@google.com as an owner
* adds mattalexander@google.com as an owner
* removes mikie@google.com as an owner
* reorders ianhua@google.com as an owner

Bug: N/A
Test: mma
Change-Id: I837b0873f1c51ce9095ddf0ad7a6085a1d23712c