Fix fully connected's reshape validation
The fully connected operation reshapes its input argument dimensions to
be compatible with its filter argument dimensions. Prior to this change,
this reshape validation would fail in the cases where input_n_elements
is less than input_size. This is because the subsequent check that
verfies input_n_elements is a multiple of input_size is skipped.

For example:
* input_n_elements = 3, input_size = 4
* batch = input_n_elements / input_size = 0
* because batch == 0, reshape validation logic is erroneously skipped

This change verifies that input_n_elements is a multiple of input_size
even in the case where input_n_elements is less than input_size.

Bug: 213595335
Test: mma
Test: NeuralNetworksTest_static
Test: libneuralnetworks_driver_fuzzer on crash case
Change-Id: I99b59abc103b24496df028eeff7c954e501fb3b5