Split validation and execution utils from OperationsUtils.h
This CL is part of an effort to move the canonical library
(neuralnetworks_types) to its own, dedicated directory. This change
splits off execution-specific utilities from OperationsUtils.{h,cpp}
because these execution utilities are not used by neuralnetworks_types,
but the rest of OperationsUtils.{h,cpp} is used by neuralnetworks_types.
These execution utilities are relocated to the new files
OperationsExecutionUtils.{h,cpp}, which are part of the module
libneuralnetworks_common. For symmetry with the execution utils, the
validation-specific parts of OperationsUtils.{h,cpp} are split into
OperationsValidationUtils.{h,cpp}.

This change also includes the following related cleanup:
* Creates the filegroup neuralnetworks_operations_validation to make the
  operation validation code available to neuralnetworks_types when it is
  relocated to a new directory
* Changes ".cpp" in #include "OperationsUtils.cpp" to ".h" in
  UtilsTest.cpp
* Changes #include "OperationResolver.h" to #include "OperationsUtils.h"
  and #include "OperationsValidationUtils.h" in Validation.cpp
* Removes some unnecesary declarations in Densify.h

Bug: 163801800
Test: mma
Test: presubmit
Change-Id: Ie715cbee11f0aa93578e17e2f49c231abb5a37d5