Clamp NNAPI HAL Version to runtime version -- runtime
The NNAPI is a Mainline Module, and its runtime version is determined by
an Android Feature Flag to remotely rollout and rollback Feature Levels
as needed. This change propagates the maximum feature level version
allowed by the runtime to the HAL utility code, and clamps the utility
code's version to the version allowed by the runtime.

Bug: N/A
Test: mma
Test: CtsNNAPITestCases
Test: NeuralNetworksTest_static
Test: locally added logging of driver version,
      setprop persist.device_config.nnapi_native.current_feature_level,
      and verified the driver version was clamped by
      current_feature_level
Change-Id: I711e43b34ba5cdfc0ff4de7d9555cee74f605220